deps: update x/tools and gopls to 842a9283 (#1031)
Includes govim changes required to satisfy the new methods the
protocol.Server interface, plus changes in response to protocol type
changes for line and column numbers.

* gopls/internal/regtest: only run in 'singleton' mode on certain GOOS 842a9283
* gopls/internal/regtest/codelens: increase the timeout on TestGCDetails bd8508e4
* internal/lsp/cache: fix AllowModfileModifications on 1.16 6d19fbfa
* internal/jsonrpc2: remove unused invalidID constant 6baea3f8
* all: upgrade all dependencies, if possible ed973575
* internal/lsp: remove redundant fields/code after source.Error deletion e7dfe027
* internal/lsp: remove the source.Error type 51ce8377
* internal/lsp/protocol/typescript: update LSP generating code 5659e493
* internal/lsp: enable semantic tokens as part of all experiments ddc05f8a
* internal/lsp: update to latest version of LSP protocol c3402e3c
* internal/lsp: stop using structured errors 2ab23861
* internal/lsp: enable -mod=readonly in workspace module mode 9b8df07b
* internal/lsp: re-enable upgrades for individual dependencies f1f686b0
* go/packages: improve go invocation errors d8a2a079
* internal/lsp/cache: remove mod upgrade code 19db92ec
* internal/lsp/protocol: use a pointer for code action's disabled field 0cef57b5

+++ cmd/govim/textedit.go
@@ -151,7 +151,7 @@ func (v *vimstate) applyProtocolTextEdits(b *types.Buffer, edits []protocol.Text
 	params := &protocol.DidChangeTextDocumentParams{
 		TextDocument: protocol.VersionedTextDocumentIdentifier{
 			TextDocumentIdentifier: b.ToTextDocumentIdentifier(),
-			Version:                float64(b.Version),
+			Version:                b.Version,
 		},
 		ContentChanges: []protocol.TextDocumentContentChangeEvent{
 			{
